# IBM-Swift/Kitura#922 Enhance features and performance of HeliumLogger