## Description
Enhance features and performance of HeliumLogger (IBM-Swift/Kitura#922)

## Motivation and Context
- Currently HeliumLogger's detailed and basic default formats do not log date/time, which should be logged by default.
- HeliumLogger formats the log message using multiple calls of string.replacingOccurrences(of). The same thing can be done much more efficiently using string interpolation or appending pre-parsed log segments.
- The default date format is "dd.MM.YYYY, HH:mm:ss". It should be in line with standard date formatting and use "yyyy-MM-dd'T'HH:mm:ss.SSS"
- Currently the check if we actually need to log based on the log level is done at the end. It should be the first thing we check
- Getting the filename from the filepath is currently done by creating a NSURL object. It is more efficient to use string.substring()
- Give users the ability to specify the time zone for the date/time

## How Has This Been Tested?
Logging a message a million times (commenting out the final print in log()) and checking the time taken:
On linux running in docker I reliably got approximately:
```
current master:
Default: 406.472931027412 seconds

This PR:
Default: 6.76490998268127 seconds
No details: 4.81400203704834 seconds
User specified: 8.68355000019073 seconds
```
On mac I reliably got approximately:
```
current master:
Default: 89.0834839940071 seconds

This PR:
Default: 10.601704955101 seconds
No details: 7.9820299744606 seconds
User specified: 14.4082700014114 seconds
```

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] I have submitted a [CLA form](https://github.com/IBM-Swift/CLA)
- [ ] If applicable, I have updated the documentation accordingly.
- [ ] If applicable, I have added tests to cover my changes.