@na-gupta 2016-12-08
This is the test code used to generate the performance numbers above. Be sure to comment out the final print() calls in HeliumLogger.swift.
```
import Foundation
import LoggerAPI

func test(_ description: String) {
    let start = Date.timeIntervalSinceReferenceDate
    for _ in 1...1000000 {
        Log.info("test")
    }
    let end = Date.timeIntervalSinceReferenceDate
    print("\(description): \(end - start) seconds")
}

let logger = HeliumLogger()
Log.logger = logger

test("Default (using interpolation)")

logger.details = false
test("No details (using interpolation)")

logger.format = "(%date) (%type) (%func) - (%msg)"
test("User specified (using replacingOccurrences)")

logger.formatter = { (date: String, type: LoggerMessageType, file: String, line: Int, function: String, msg: String) in
    return "\(date) \(type) \(function) - \(msg)"
}
test("User specified (using interpolation)")
```