@ianpartridge 2016-12-08
This is really hard to review when it's 1 commit containing performance improvements, functional changes, new API, and bugfixes.

It would be much easier to review as a series of commits, each with one change, building on the previous one. Or even better, as multiple PRs.

Just merging this as 1 commit makes it hard to understand the history of HeliumLogger and means if any part contains a regression it will be harder to back that particular piece out, rather than a simple `git revert` of one commit.

As general practice, I don't think we should merge PRs like this.

@shmuelk what is your opinion?