@shmuelk 2016-12-08
I kind of agree with Ian that is an awful lot of change here all at once. I am also not quite so sure we needed that much API change to get better performance.

We need to discuss and think about this some more.