@na-gupta 2016-12-08
The vast majority of the code changes in this PR are related to performance.

But there's just one API addition: `formatter` related to performance. There's one more API addition: `timeZone` to allow users to specify the time zone for date/time logging.

There are no API deprecations or any other API changes.