@na-gupta 2016-12-08
I've updated this to construct the log entry by appending log segments instead of replacingOccurrences. Performance is very close to string interpolation removing need for user specified formatting closure.

So now the only new API are `timeZone: TimeZone?` to allow users to specify the time zone for date/time logging and `fullFilePath: Bool` to allow users to log the full file path if they prefer.