@shmuelk 2016-12-13
Looks very good. I really liked the pre-parsing of the custom log format.