@na-gupta 2017-01-11
@djones6  The 1 NSURL creation per `log()` call was a relatively small part of the performance gain in my tests. The bulk of the performance came from replacing the multiple invocations of `message.replacingOccurrences(of: stringValue, with: replaceValue)` with `line.append(value)`.